Use FieldFE in GenericField.
- replace FieldElementwise

+++ src/mesh/partitioning.cc
@@ -234,13 +234,13 @@ void Partitioning::id_maps(int n_ids, LongIdx *id_4_old,  Distribution * &new_ds
 
 
 
-shared_ptr< vector<double> > Partitioning::subdomain_id_field_data() {
+shared_ptr< vector<int> > Partitioning::subdomain_id_field_data() {
 	OLD_ASSERT(loc_part_, "Partition is not yet computed.\n");
     if (!seq_part_) {
     	unsigned int seq_size=(init_el_ds_->myp() == 0) ? init_el_ds_->size() : 1;
     	//seq_part_.resize(seq_size);
-    	seq_part_ = make_shared< vector<double> >(seq_size);
-        std::vector<double> &vec = *( seq_part_.get() );
+    	seq_part_ = make_shared< vector<int> >(seq_size);
+        std::vector<int> &vec = *( seq_part_.get() );
 
         MPI_Gatherv(loc_part_, init_el_ds_->lsize(), MPI_INT,
                 &vec[0],
