+++ src/mesh/partitioning.hh
@@ -81,7 +81,7 @@ public:
     /**
      * Creates and returns vector with element partitioning for output.
      */
-    shared_ptr< vector<double> > subdomain_id_field_data();
+    shared_ptr< vector<int> > subdomain_id_field_data();
 
     /**
      * Obsolete see source file for doc.
@@ -127,7 +127,7 @@ private:
     /// Original distribution of elements. Depends on type of partitioner
     Distribution *init_el_ds_;
     /// Sequential partitioning for output.
-    shared_ptr< vector<double> > seq_part_;
+    shared_ptr< vector<int> > seq_part_;
 
     /**
      * Creates sparse parallel graph from the mesh (using algorithm given by the key "graph_type" of the input record accessor @p in_
